Fix: Failing to create template
The secret provided via argument (`-e` flag) [1] was losing the value before
listing the tools. This caused failures to connect to the MCP server.

The strategy now is to replace the secret value only when encoding the object
to generate the template.

There was also a missing `--load` flag when executing `docker buildx build` to
correctly load the image into the memory.

[1] https://github.com/rafaeljusto/mcp-registry/blob/17fdb6eb0a4b47d4ac1a33535554a6fa7e2b2dd3/cmd/create/main.go#L177-L181